requirements.txt
Cython==0.28.2
ConfigSpace==0.4.6
numpy==1.14.2
pandas==0.22.0
scikit-learn==0.19.1
scipy==1.0.1
joblib
auto-sklearn==0.4.0
psycopg2-binary
rpy2
tzlocal
pygam